feat(libcoaligned): bake the enum-drift engine into the invariant kit
Lift the enumeration-drift engine out of the repo-local .coaligned/invariants/
and into libcoaligned as a reusable build-kit primitive, so a consuming repo's
rule module carries only policy (its topics registry) and a one-line delegation.

- src/enum-drift-grammar.js: the source probes (fs-glob, md-table), list/count
  extractors, and consumer parser, with filesystem access parameterized out
  (passed as fsSync) to stay clean under ambient-deps.
- src/enum-drift.js: the build/seed orchestration + ENUM_DRIFT_RULES, taking a
  parsed topics registry; re-exports the grammar.
- invariant-kit.js: inject kit.enumDrift.{build,seed} into the build kit and
  enumDriftRules into the rule kit, beside scan/grep/restatementDrift. Hints
  generalized so they read correctly in any consuming repo.

Support a bare comma-separated run of inline code spans as a list shape in
extractList (e.g. `a`, `b`, `c`), guarded so it fires only when the whole span
is code spans plus separators — prose/description commas still decline.

Slim .coaligned/invariants/enumeration-drift.rules.mjs to 18 lines of policy
delegating to the kit; delete lib/enum-drift.mjs; topics.yml is unchanged. Tests
re-point to libcoaligned/src, split by behaviour family under the line cap with
shared setup in test/helpers.js, and cover the new shape. Also carries the
CLAUDE.md comma-separated enum form (now valid) and the KATA.md bullet reflow.

Additive libcoaligned API (new kit members) — minor version bump at release.
